// src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export interface Scalars {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSONObject: { input: any; output: any; }
  Time: { input: any; output: any; }
}
// ...
export interface Host {
  deviceType?: Maybe<Scalars['String']['output']>;
  /** The host field contains the "hostname" in Zabbix */
  host: Scalars['String']['output'];
  hostgroups?: Maybe<Array<HostGroup>>;
  hostid: Scalars['ID']['output'];
  name?: Maybe<Scalars['String']['output']>;
  tags?: Maybe<Scalars['JSONObject']['output']>;
}

export interface HostGroup {
  __typename?: 'HostGroup';
  groupid: Scalars['ID']['output'];
  name?: Maybe<Scalars['String']['output']>;
}
// ...
export interface Inventory {
  __typename?: 'Inventory';
  location?: Maybe<Location>;
}
// ...
export interface Location extends GpsPosition {
  __typename?: 'Location';
  latitude?: Maybe<Scalars['Float']['output']>;
  longitude?: Maybe<Scalars['Float']['output']>;
  name?: Maybe<Scalars['String']['output']>;
}
// ...
export interface Template {
  __typename?: 'Template';
  name?: Maybe<Scalars['String']['output']>;
  templateid: Scalars['String']['output'];
}
// ...
export interface ZabbixHost extends Host {
  __typename?: 'ZabbixHost';
  deviceType?: Maybe<Scalars['String']['output']>;
  host: Scalars['String']['output'];
  hostgroups?: Maybe<Array<HostGroup>>;
  hostid: Scalars['ID']['output'];
  inventory?: Maybe<Inventory>;
  items?: Maybe<Array<ZabbixItem>>;
  name?: Maybe<Scalars['String']['output']>;
  parentTemplates?: Maybe<Array<Template>>;
  tags?: Maybe<Scalars['JSONObject']['output']>;
}

export interface ZabbixItem {
  __typename?: 'ZabbixItem';
  attributeName?: Maybe<Scalars['String']['output']>;
  deviceType?: Maybe<Scalars['String']['output']>;
  hostid?: Maybe<Scalars['Int']['output']>;
  hosts?: Maybe<Array<Maybe<Host>>>;
  itemid: Scalars['Int']['output'];
  key_: Scalars['String']['output'];
  lastclock?: Maybe<Scalars['Int']['output']>;
  lastvalue?: Maybe<Scalars['String']['output']>;
  name: Scalars['String']['output'];
  status?: Maybe<DeviceStatus>;
  topicType?: Maybe<Scalars['String']['output']>;
  type?: Maybe<DeviceCommunicationType>;
  value_type: Scalars['Int']['output'];
}